requirements.txt
Adafruit-DHT==1.4.0
Django==2.2.10
pytz==2020.1
RPi.GPIO==0.7.0
sqlparse==0.3.1
